refactor: sealed classes for irrationals
